refactor: move attribute and valueType logic to lib
