+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-03-06T15:14:09.076Z\n"
-"PO-Revision-Date: 2025-03-06T15:14:09.077Z\n"
+"POT-Creation-Date: 2025-03-12T16:02:01.038Z\n"
+"PO-Revision-Date: 2025-03-12T16:02:01.038Z\n"
 
 msgid "schemas"
 msgstr "schemas"
@@ -93,26 +93,11 @@ msgstr "Filter options"
 msgid "Retry"
 msgstr "Retry"
 
+msgid "Required"
+msgstr "Required"
+
 msgid "Please enter a GeoJSON value"
 msgstr "Please enter a GeoJSON value"
-
-msgid "Please provide a positive integer"
-msgstr "Please provide a positive integer"
-
-msgid "Please provide a negative integer"
-msgstr "Please provide a negative integer"
-
-msgid "Please provide a 0 or positive integer"
-msgstr "Please provide a 0 or positive integer"
-
-msgid "Please provide valid percantage (0-100)."
-msgstr "Please provide valid percantage (0-100)."
-
-msgid "Please provide valid unit interval (0-1)."
-msgstr "Please provide valid unit interval (0-1)."
-
-msgid "Must be checked"
-msgstr "Must be checked"
 
 msgid "Save and close"
 msgstr "Save and close"
@@ -135,6 +120,9 @@ msgstr "Something went wrong when submitting the form"
 msgid "<No value>"
 msgstr "<No value>"
 
+msgid "Please enter a GeoJSON value."
+msgstr "Please enter a GeoJSON value."
+
 msgid "Attributes"
 msgstr "Attributes"
 
@@ -143,9 +131,6 @@ msgstr "Set up information for the attributes assigned to {{modelName}}."
 
 msgid "Code"
 msgstr "Code"
-
-msgid "Required"
-msgstr "Required"
 
 msgid "Description"
 msgstr "Description"
@@ -1095,11 +1080,26 @@ msgstr "Cannot save empty object"
 msgid "Created successfully"
 msgstr "Created successfully"
 
+msgid "Please provide a positive integer"
+msgstr "Please provide a positive integer"
+
+msgid "Please provide a negative integer"
+msgstr "Please provide a negative integer"
+
+msgid "Please provide a 0 or positive integer"
+msgstr "Please provide a 0 or positive integer"
+
+msgid "Please provide valid percantage (0-100)."
+msgstr "Please provide valid percantage (0-100)."
+
+msgid "Please provide valid unit interval (0-1)."
+msgstr "Please provide valid unit interval (0-1)."
+
+msgid "Must be checked"
+msgstr "Must be checked"
+
 msgid "Period type"
 msgstr "Period type"
-
-msgid "Level"
-msgstr "Level"
 
 msgid "Offline levels"
 msgstr "Offline levels"
