+++ src/components/fields/valueType/index.ts
@@ -1,2 +1,1 @@
 export * from './FormFieldByValueType'
-export * from './validateByValueType'
