+++ src/components/form/attributes/index.ts
@@ -1,2 +1,1 @@
 export { CustomAttributesSection } from './CustomAttributes'
-export { useCustomAttributesQuery } from './useCustomAttributesQuery'
