+++ src/lib/models/index.ts
@@ -6,3 +6,5 @@ export { useIsFieldValueUnique } from './useIsFieldValueUnique'
 export { useCheckMaxLengthFromSchema } from './useCheckMaxLengthFromSchema'
 export * from './useModelMultiSelectQuery'
 export * from './usePaginatedModelQuery'
+export * from './attributes'
+export * from './valueType'
